backend/requirements.txt
flask
flask-cors
google-generativeai
pandas
numpy
scikit-learn
